доработки bff post

+++ readme/apps/blogs/src/app/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import {HttpException, HttpStatus, Injectable} from '@nestjs/common';
 import {BlogPostRepository} from '../blog-post/blog-post.repository';
 import {BlogPostEntity} from '../blog-post/blog-post.entity';
 import {CreateVideoDto} from './dto/create-video.dto';
@@ -9,6 +9,8 @@ import {CreateLinkDto} from './dto/create-link.dto';
 import {CreateRepostDto} from './dto/create-repost.dto';
 import {PostQuery} from './query/post.query';
 import {Post} from '@readme/shared-types';
+import {UpdateDto} from './dto/update.dto';
+import {POST_DELETE_FORBIDDEN, POST_UPDATE_FORBIDDEN} from './post.constant';
 
 @Injectable()
 export class PostService {
@@ -30,19 +32,33 @@ export class PostService {
   }
 
   async createRepost(dto: CreateRepostDto ) {
-    const {idOriginal} = dto;
+    const {idOriginal, author} = dto;
     const postOriginal = await this.blogPostRepository.findById(+idOriginal);
     const postNew ={
       ...postOriginal
-      ,author: ''
+      ,author: author
       ,isRepost: true
     };
     const postEntity = new BlogPostEntity(postNew);
     return this.blogPostRepository.create(postEntity);
   }
 
-  async deletePost(id: number): Promise<void> {
-    this.blogPostRepository.destroy(id);
+  async updateAndValidate(id: number, dto: UpdateDto) {
+    const {author} = dto;
+    const postOriginal = await this.blogPostRepository.findById(id);
+    if (postOriginal.author !== author) {
+      throw new HttpException(POST_UPDATE_FORBIDDEN, HttpStatus.FORBIDDEN);
+    }
+    const postEntity = new BlogPostEntity({...dto});
+    return await this.blogPostRepository.update(id, postEntity);
+  }
+
+  async deleteAndValidate(id: number, author: string): Promise<void> {
+    const postOriginal = await this.blogPostRepository.findById(id);
+    if (postOriginal.author !== author) {
+      throw new HttpException(POST_DELETE_FORBIDDEN, HttpStatus.FORBIDDEN);
+    }
+    return await this.blogPostRepository.destroy(id);
   }
